requirements.txt
torch>=1.10.0
torchvision>=0.11.1
transformers>=4.18.0
datasets>=2.0.0
pillow>=8.4.0
requests>=2.27.1
nltk>=3.6.7
tqdm>=4.62.3
matplotlib>=3.5.1
scikit-learn>=1.0.2
spacy>=3.0.0 